update oink rec verifier

+++ barretenberg/cpp/src/barretenberg/ultra_honk/oink_verifier.cpp
@@ -37,9 +37,9 @@ template <IsUltraFlavor Flavor> void OinkVerifier<Flavor>::verify()
 template <IsUltraFlavor Flavor> void OinkVerifier<Flavor>::execute_preamble_round()
 {
     // TODO(Adrian): Change the initialization of the transcript to take the VK hash?
-    const uint32_t circuit_size = verification_key->verification_key->circuit_size;
-    const uint32_t public_input_size = verification_key->verification_key->num_public_inputs;
-    const uint32_t pub_inputs_offset = verification_key->verification_key->pub_inputs_offset;
+    const uint64_t circuit_size = verification_key->verification_key->circuit_size;
+    const uint64_t public_input_size = verification_key->verification_key->num_public_inputs;
+    const uint64_t pub_inputs_offset = verification_key->verification_key->pub_inputs_offset;
 
     transcript->add_to_hash_buffer(domain_separator + "circuit_size", circuit_size);
     transcript->add_to_hash_buffer(domain_separator + "public_input_size", public_input_size);
